TBD: platform/x86: add Intel PMU driver
-- %< --
The proposed driver is a minimum stub to disable PMU and turn all devices on.
This allows to use serial buses and other controllers, though the power
consumption is at maximum.

The driver should evolve to a proper one at some point where actual work is
done behind runtime PM.
-- %< --

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>